requirements.txt
appnope==0.1.4
arrow==1.3.0
asttokens==3.0.0
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.8
idna==3.10
importlib_metadata==8.6.1
ipyevents==2.0.2
ipykernel==6.29.5
ipympl==0.9.6
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mne==1.9.0
nest_asyncio==1.6.0
nibabel==5.3.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.1
pyvista==0.44.2
pyvistaqt==0.11.2
PyYAML==6.0.2
pyzmq==26.2.0
QtPy==2.4.3
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
scooby==0.10.0
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
stack_data==0.6.3
text-unidecode==1.3
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vtk==9.3.1
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.13
zipp==3.21.0
